change button to follow/unfollow. add following indicator. add datastore to persist following user ids

+++ app/src/main/java/com/example/stackoverflowusers/data/User.kt
@@ -3,7 +3,7 @@ package com.example.stackoverflowusers.data
 import com.google.gson.annotations.SerializedName
 
 data class User(
-    @SerializedName("user_id") val userId: Long,
+    @SerializedName("user_id") val userId: Int,
     @SerializedName("display_name") val displayName: String,
     val reputation: Int,
     @SerializedName("profile_image") val profileImage: String?,
